+++ src/vue/fields/InputField.vue
@@ -7,6 +7,7 @@
       'input-field--readonly': $attrs.readonly || $attrs.readonly === '',
       'input-field--disabled': $attrs.disabled,
       'input-field--pwd-toggle-present': isPasswordType,
+      'input-field--phone-number': isPhoneNumberType,
     }"
   >
     <input
@@ -14,7 +15,8 @@
       v-on="listeners"
       class="input-field__input"
       :class="{
-        'input-field__input--autofill-white': whiteAutofill
+        'input-field__input--autofill-white': whiteAutofill,
+        'input-field__phone-number-input': isPhoneNumberType
       }"
       :type="isPasswordType && isPasswordShown ? 'text' : type"
       :value="value"
@@ -42,7 +44,12 @@
       />
     </button>
 
-    <span class="input-field__label">
+    <span
+      class="input-field__label"
+      :class="{
+        'input-field__label--phone-number': isPhoneNumberType,
+      }"
+    >
       {{ label }}
 
       <template v-if="isCapsLockOn">
@@ -63,6 +70,13 @@ import { MathUtil } from '@/js/utils'
 
 const EVENTS = {
   input: 'input',
+}
+
+const INPUT_TYPES = {
+  phoneNumber: 'phone-number',
+  password: 'password',
+  number: 'number',
+
 }
 
 export default {
@@ -92,7 +106,11 @@ export default {
     },
 
     isPasswordType () {
-      return this.type === 'password'
+      return this.type === INPUT_TYPES.password
+    },
+
+    isPhoneNumberType () {
+      return this.type === INPUT_TYPES.phoneNumber
     },
   },
 
@@ -132,7 +150,7 @@ export default {
     },
 
     normalizeTargetValue (target) {
-      if (this.type === 'number' && target.value !== '') {
+      if (this.type === INPUT_TYPES.number && target.value !== '') {
         target.value = this.normalizeDecimalPrecision(
           this.normalizeRange(target.value)
         )
@@ -404,5 +422,26 @@ $pwd-toggle-btn-width: 3.2rem;
     margin-top: $field-error-margin-top;
     overflow: hidden;
   }
+}
+
+.input-field--phone-number:before {
+  position: absolute;
+  top: 1.4rem;
+  content: '\002B';
+  left: 0;
+  font-size: 1.8rem;
+  line-height: 1.25;
+}
+
+.input-field__label--phone-number {
+  .input-field__input:not(:focus):placeholder-shown ~ & {
+    top: 0;
+
+    @include label-font-sizes;
+  }
+}
+
+.input-field__phone-number-input {
+  padding: 1.5rem 0 0.6rem 1rem;
 }
 </style>